refactor(nx-cmake): refactor graph reduction utils, add test suites for refactors
