+++ packages/nx-cmake/src/utils/generatorUtils/getLanguageExtension/getLanguageExtension.ts
@@ -1,3 +1,3 @@
-import type { C } from '../../../models/types';
+import type { C, CTag } from '../../../models/types';
 
-export const getLanguageExtension = (l: C) => (l === 'C' ? 'c' : 'cpp');
+export const getLanguageExtension = (l: C): CTag => (l === 'C' ? 'c' : 'cpp');
