+++ packages/nx-cmake/src/utils/generatorUtils/getProjectType/getProjectTypeFromConfigFileContent/getProjectFileFromConfigFileContent.ts
@@ -1,0 +1,16 @@
+import { CProjectType } from '../../../../models/types';
+
+export const getProjectTypeFromConfigFileContent = (
+    content: string
+): CProjectType => {
+    if (content.includes('enable_testing()')) {
+        return CProjectType.Test;
+    }
+    if (content.includes('set_library_settings')) {
+        return CProjectType.Lib;
+    }
+    if (content.includes('set_binary_settings')) {
+        return CProjectType.App;
+    }
+    throw new Error('Failed to determine project type');
+};
