+++ packages/nx-cmake/src/utils/generatorUtils/getProjectType/readProjectFile/readProjectFile.ts
@@ -1,0 +1,12 @@
+import { readFileSync } from 'fs';
+import { PROJECT_FILE } from '../../../../config/projectFilePattern';
+
+export const readProjectFile = (projectFile: string): string => {
+    if (!projectFile.includes(PROJECT_FILE)) {
+        throw new Error('Invalid project file');
+    }
+    const fileContent = readFileSync(projectFile, {
+        encoding: 'utf-8',
+    });
+    return fileContent;
+};
