+++ packages/nx-cmake/src/utils/graphUtils/reduceDependenciesTransitively/buildAdjacencyList/buildAdjacencyList.spec.ts
@@ -1,0 +1,6 @@
+import type { ProjectGraphDependencyWithFile } from '@nx/devkit';
+import { buildAdjacencyList } from './buildAdjacencyList';
+
+describe('buildAdjacencyList', () => {
+    it.todo('should build adjacency list');
+});
