+++ packages/nx-cmake/src/utils/graphUtils/reduceDependenciesTransitively/buildGraphFromDeps/buildGraphFromDeps.spec.ts
@@ -1,0 +1,7 @@
+import type { Graph } from '../../../../models/types';
+import type { ProjectGraphDependencyWithFile } from '@nx/devkit';
+import { buildGraphFromDeps } from './buildGraphFromDeps';
+
+describe('buildGraphFromDeps', () => {
+    it.todo('should build graph from deps');
+});
