+++ packages/nx-cmake/src/utils/graphUtils/reduceDependenciesTransitively/getTransitiveDependencies/getTransitiveDependencies.spec.ts
@@ -1,0 +1,6 @@
+import type { Graph } from '../../../../models/types';
+import { getTransitiveDependencies } from './getTransitiveDependencies';
+
+describe('getTransitiveDependencies', () => {
+    it.todo('should get transitive dependencies');
+});
